+++ Holiday_For_Children.csv
@@ -1,0 +1,6 @@
+price, duration, childrenNumber, ageCategory, MasterClassType
+400.0, 1.3, 5, 16, CLAY_MOLDING
+price, duration, childrenNumber, ageCategory, AnimatorType, animatorsQuantity
+300.0, 2.3, 5, 8, PHAMTOMIMIST, 2
+price, duration, childrenNumber, ageCategory, QuestLocation, QuestDifficulty
+800.0, 4.0, 8, 14, CITY_CENTRE, NORMAL
