--- main/java/ua/lviv/iot/lab_3/AnimatorType.java
@@ -1,7 +1,0 @@
-package ua.lviv.iot.lab_3;
-
-public enum AnimatorType {
-
-    CLOWN, MAGICIAN, PHAMTOMIMIST, HERO_COSTUME;
-
-}
