--- main/java/ua/lviv/iot/lab_3/MasterClassType.java
@@ -1,7 +1,0 @@
-package ua.lviv.iot.lab_3;
-
-public enum MasterClassType {
-
-    PAINTING, APPLIQUE, CLAY_MOLDING;
-
-}
