--- main/java/ua/lviv/iot/lab_3/QuestDifficulty.java
@@ -1,7 +1,0 @@
-package ua.lviv.iot.lab_3;
-
-public enum QuestDifficulty {
-
-    EASY, NORMAL, HARD;
-
-}
