--- main/java/ua/lviv/iot/lab_3/QuestLocation.java
@@ -1,7 +1,0 @@
-package ua.lviv.iot.lab_3;
-
-public enum QuestLocation {
-
-    CITY_CENTRE, PARK, INFOORS;
-
-}
