+++ src/main/java/ua/lviv/iot/lab_3/managers/HolidayManagerInterface.java
