+++ src/main/java/ua/lviv/iot/lab_3/models/AnimatorType.java
@@ -1,0 +1,7 @@
+package ua.lviv.iot.lab_3.models;
+
+public enum AnimatorType {
+
+    CLOWN, MAGICIAN, PHAMTOMIMIST, HERO_COSTUME;
+
+}
