+++ src/main/java/ua/lviv/iot/lab_3/models/MasterClassType.java
@@ -1,0 +1,7 @@
+package ua.lviv.iot.lab_3.models;
+
+public enum MasterClassType {
+
+    PAINTING, APPLIQUE, CLAY_MOLDING;
+
+}
