+++ src/main/java/ua/lviv/iot/lab_3/models/Quest.java
@@ -1,0 +1,41 @@
+package ua.lviv.iot.lab_3.models;
+
+public class Quest extends HolidayForChildren {
+
+    private QuestLocation location;
+    private QuestDifficulty difficulty;
+
+    public Quest() {
+
+    }
+
+    public Quest(final double price,
+                 final double duration,
+                 final int childrenNumber,
+                 final int ageCategory,
+                 final QuestLocation location,
+                 final QuestDifficulty difficulty) {
+        super(price, duration, childrenNumber, ageCategory);
+        this.location = location;
+        this.difficulty = difficulty;
+    }
+
+    @Override
+    public String getHeaders() {
+        return super.getHeaders() + ", QuestLocation, QuestDifficulty";
+    }
+
+    @Override
+    public String toCSV() {
+        return super.toCSV() + ", " + getLocation() + ", " + getDifficulty();
+    }
+
+    public final QuestLocation getLocation() {
+        return location;
+    }
+
+    public final QuestDifficulty getDifficulty() {
+        return difficulty;
+    }
+
+}
