+++ src/main/java/ua/lviv/iot/lab_3/models/QuestDifficulty.java
@@ -1,0 +1,7 @@
+package ua.lviv.iot.lab_3.models;
+
+public enum QuestDifficulty {
+
+    EASY, NORMAL, HARD;
+
+}
