+++ src/main/java/ua/lviv/iot/lab_3/models/QuestLocation.java
@@ -1,0 +1,7 @@
+package ua.lviv.iot.lab_3.models;
+
+public enum QuestLocation {
+
+    CITY_CENTRE, PARK, INFOORS;
+
+}
